backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "absorpig",
    "fastapi[standard]>=0.115.3",
    "pandas-stubs>=2.2.3.241009",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
    "requests>=2.32.3",
]

[tool.uv.sources]
absorpig = { path = "pkg/absorpig-0.1.0-py3-none-any.whl" }
